+++ packages/notification/channel/dingtalk.ts
@@ -25,7 +25,7 @@
 /* eslint-disable camelcase */
 import format from 'date-fns/format'
 import Joi from 'joi'
-import type { NotificationMessage } from '..'
+import type { NotificationMessage } from '.'
 import { sendHttpRequest } from '../utils/http'
 
 type NotificationData = {
