+++ packages/notification/channel/pushbullet.ts
@@ -24,7 +24,7 @@
 
 import Joi from 'joi'
 import { sendHttpRequest } from '../utils/http'
-import type { NotificationMessage } from '..'
+import type { NotificationMessage } from '.'
 
 type NotificationData = {
   token: string
