+++ packages/notification/channel/smtp.ts
@@ -26,7 +26,7 @@ import * as nodemailer from 'nodemailer'
 import Mailgen from 'mailgen'
 import Joi from 'joi'
 
-import type { NotificationMessage } from '..'
+import type { NotificationMessage } from '.'
 
 type NotificationData = {
   hostname: string
