+++ packages/notification/channel/webhook.ts
@@ -23,7 +23,7 @@
  **********************************************************************************/
 
 import Joi from 'joi'
-import type { NotificationMessage } from '..'
+import type { NotificationMessage } from '.'
 import { sendHttpRequest } from '../utils/http'
 
 type NotificationData = {
