+++ packages/notification/channel/whatsapp.ts
@@ -23,7 +23,7 @@
  **********************************************************************************/
 
 import Joi from 'joi'
-import type { NotificationMessage } from '..'
+import type { NotificationMessage } from '.'
 import { authorize } from '../utils/authorization'
 import { sendHttpRequest } from '../utils/http'
