+++ packages/notification/validator/notification.ts
@@ -22,7 +22,7 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import { channels, type Notification } from '..'
+import { channels, type Notification } from '../channel'
 
 export const validateNotification = async (
   notifications: Notification[]
