+++ src/components/logger/startup-message.ts
@@ -27,8 +27,8 @@ import isUrl from 'is-url'
 import boxen from 'boxen'
 import chalk from 'chalk'
 import type { Config } from '../../interfaces/config'
-import type { Notification } from '@hyperjumptech/monika-notification'
-import { channels } from '@hyperjumptech/monika-notification'
+import type { Notification } from '@hyperjumptech/monika-notification/channel'
+import { channels } from '@hyperjumptech/monika-notification/channel'
 import type { Probe, ProbeAlert } from '../../interfaces/probe'
 import type { RequestConfig } from '../../interfaces/request'
 import { log } from '../../utils/pino'
