+++ src/jobs/summary-notification.ts
@@ -25,7 +25,7 @@ import { hostname } from 'os'
 import format from 'date-fns/format'
 import { getConfig } from '../components/config'
 import { getSummary } from '../components/logger/history'
-import { sendNotifications } from '../components/notification'
+import { sendNotifications } from '@hyperjumptech/monika-notification'
 import {
   getOSName,
   getMonikaInstance,
@@ -48,7 +48,7 @@ import events from '../events'
 import { Config } from '../interfaces/config'
 
 import { getEventEmitter } from '../utils/events'
-import type { Notification } from '@hyperjumptech/monika-notification'
+import type { Notification } from '@hyperjumptech/monika-notification/channel'
 import type { Probe } from '../interfaces/probe'
 const eventEmitter = getEventEmitter()
